CHANGE #7
1. entities.py
"Simplification" in method "move()" for "player" class.
"", because a teacher told me that my code is hard to read.

2. functionalObjects.py
    > Class "screen": Change in the place of methods.
    > Class "pointer": Method "select_button()" finished; method "move()" created.
    > Class "button": Method "selected()" created.

3. main.py
Preparing a functional home menu screen.

4. screen.py
Changed the black tiles (⬛️) in the list "maze" for double blank spaces ("  ").
A second list ("original_maze_with_black_tiles") was created as backup to develop new mazes.